do_cmd_read_scroll,do_cmd_read_scroll_aux関数にリファクタリングによる引数追加。
git-svn-id: http://svn.osdn.net/svnroot/deeangband@442 89dd94fa-70ca-40ae-8747-2094edff379b

+++ v0_1_0/trunk/src/wild.c
@@ -678,8 +678,8 @@ void wilderness_gen_small()
 	panel_col_min = cur_wid;
 
 	/* Place the player */
-	p_ptr->fx = p_ptr->wilderness_x;
-	p_ptr->fy = p_ptr->wilderness_y;
+	p_ptr->fx = (byte)p_ptr->wilderness_x;
+	p_ptr->fy = (byte)p_ptr->wilderness_y;
 
 	p_ptr->town_num = 0;
 }
